Refactor random-id.tsx to randomId.ts
