+++ designer/client/reducers/listReducer.tsx
@@ -1,4 +1,4 @@
-import randomId from "../random-id";
+import randomId from "../randomId";
 import React, { createContext, useContext, useReducer } from "react";
 import { DataContext } from "../context";
 import { ListActions } from "./listActions";
